feat: tokens and user validations

+++ server/app/routes/userRouter.py
@@ -2,9 +2,57 @@ from flask import Blueprint, request, jsonify
 
 from ..repositories.UserRepository import UserRepository
 from ..errors.AppError import AppError
+from ..validations.userValidation import validateData, emailRegex, passwordRegex
+from ..validations.tokenValidation import generateToken, verifyToken, private_key
 
 userRouter = Blueprint("user-routes", __name__, url_prefix="/user")
 userRepository = UserRepository()
+
+token = None
+
+@userRouter.route("/login", methods=['GET'])
+def loginUser():
+    data = request.get_json()
+    if not validateData("email", "password", data=data):
+        raise AppError("Invalid body or data", 400)
+
+    user = userRepository.login(data["email"], data["password"])
+    global token
+    token = generateToken(data, private_key)
+    return jsonify(user.getAttributes())
+
+
+@userRouter.route("/logout/<userId>", methods=['GET'])
+def logoutUser(userId):
+    data = request.get_json()
+    if not validateData("password", data=data):
+        raise AppError("Invalid body or data", 400)
+
+    global token
+    if not verifyToken(token, private_key):
+        raise AppError("Unauthorized", 401)
+
+    user = userRepository.logout(userId, data["password"])
+    token = None
+    return jsonify(user.getAttributes())
+
+
+@userRouter.route("/create", methods=['POST'])
+def createUser():
+    data = request.get_json()
+
+    if not validateData("name", "email", "password", data=data):
+        raise AppError("Invalid body or data", 400)
+    if not emailRegex(data["email"]):
+        raise AppError("Invalide email", 400)
+    if not passwordRegex(data["password"]):
+        raise AppError("Invalid password", 400)
+
+    newUser = userRepository.create(data)
+    global token
+    token = generateToken(data, private_key)
+    return jsonify(newUser)
+
 
 @userRouter.route("/list", methods=['GET'])
 def listUsers():
@@ -18,23 +66,13 @@ def showUser(userId):
     return jsonify(user)
 
 
-@userRouter.route("/create", methods=['POST'])
-def createUser():
-    data = request.get_json()
-    if data == None:
-        raise AppError("No body", 400)
-    elif "name" not in data or "email" not in data or "password" not in data: # ?
-        raise AppError("Ivalid body", 400)
-
-    newUser = userRepository.create(data)
-    return jsonify(newUser)
-
-
 @userRouter.route("/update/<userId>", methods=['PUT'])
 def updateUser(userId):
     data = request.get_json()
-    if data == None:
-        raise AppError("No body", 400)
+
+    global token
+    if not verifyToken(token, private_key):
+        raise AppError("Unauthorized", 401)
 
     updatedUser = userRepository.update(userId, data)
     return jsonify(updatedUser)
@@ -42,5 +80,9 @@ def updateUser(userId):
 
 @userRouter.route("/delete/<userId>", methods=['DELETE'])
 def deleteUser(userId):
+    global token
+    if not verifyToken(token, private_key):
+        raise AppError("Unauthorized", 401)
+
     user = userRepository.delete(userId)
     return jsonify(user)
